Change package from org.duchastel to com.duchastel
