+++ composeApp/src/androidMain/kotlin/com/duchastel/simon/solenne/parcel/Parcelize.android.kt
@@ -1,0 +1,8 @@
+package com.duchastel.simon.solenne.parcel
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+actual typealias Parcelize = Parcelize
+
+actual typealias Parcelable = Parcelable
