--- composeApp/src/androidMain/kotlin/org/duchastel/simon/solenne/parcel/Parcelize.android.kt
@@ -1,8 +1,0 @@
-package org.duchastel.simon.solenne.parcel
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-actual typealias Parcelize = Parcelize
-
-actual typealias Parcelable = Parcelable
