+++ composeApp/src/commonMain/kotlin/com/duchastel/simon/solenne/parcel/Parcelize.kt
@@ -1,0 +1,10 @@
+package com.duchastel.simon.solenne.parcel
+
+@OptIn(ExperimentalMultiplatform::class)
+@OptionalExpectation
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.BINARY)
+expect annotation class Parcelize()
+
+// necessary to shim the Android Parcelable class for Circuit
+expect interface Parcelable
