+++ composeApp/src/commonMain/kotlin/com/duchastel/simon/solenne/screens/chat/ChatPresenter.kt
@@ -1,0 +1,11 @@
+package com.duchastel.simon.solenne.screens.chat
+
+import androidx.compose.runtime.Composable
+import com.slack.circuit.runtime.presenter.Presenter
+
+class ChatPresenter : Presenter<ChatScreen.State> {
+    @Composable
+    override fun present(): ChatScreen.State {
+        return ChatScreen.State()
+    }
+}
