+++ composeApp/src/commonMain/kotlin/com/duchastel/simon/solenne/screens/chat/ChatScreen.kt
@@ -1,0 +1,15 @@
+package com.duchastel.simon.solenne.screens.chat
+
+import com.duchastel.simon.solenne.parcel.Parcelize
+import com.slack.circuit.runtime.CircuitUiEvent
+import com.slack.circuit.runtime.CircuitUiState
+import com.slack.circuit.runtime.screen.Screen
+
+@Parcelize
+data object ChatScreen: Screen {
+    data class State(
+        val eventSink: (Event) -> Unit = {},
+    ): CircuitUiState
+
+    sealed interface Event : CircuitUiEvent
+}
