+++ composeApp/src/desktopMain/kotlin/com/duchastel/simon/solenne/parcel/Parcelize.desktop.kt
@@ -1,0 +1,3 @@
+package com.duchastel.simon.solenne.parcel
+
+actual interface Parcelable
