+++ composeApp/src/iosMain/kotlin/com/duchastel/simon/solenne/MainViewController.kt
@@ -1,0 +1,8 @@
+package com.duchastel.simon.solenne
+
+import androidx.compose.material.Text
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController {
+    Text("iOS is not yet supported")
+}
