--- composeApp/src/iosMain/kotlin/org/duchastel/simon/solenne/MainViewController.kt
@@ -1,8 +1,0 @@
-package org.duchastel.simon.solenne
-
-import androidx.compose.material.Text
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController {
-    Text("iOS is not yet supported")
-}
