+++ composeApp/src/nativeMain/kotlin/com/duchastel/simon/solenne/parcel/Parcelize.native.kt
@@ -1,0 +1,3 @@
+package com.duchastel.simon.solenne.parcel
+
+actual interface Parcelable
