--- composeApp/src/nativeMain/kotlin/org/duchastel/simon/solenne/parcel/Parcelize.native.kt
@@ -1,3 +1,0 @@
-package org.duchastel.simon.solenne.parcel
-
-actual interface Parcelable
