+++ composeApp/src/wasmJsMain/kotlin/com/duchastel/simon/solenne/parcel/Parcelize.wasmJs.kt
@@ -1,0 +1,3 @@
+package com.duchastel.simon.solenne.parcel
+
+actual interface Parcelable
