--- composeApp/src/wasmJsMain/kotlin/org/duchastel/simon/solenne/parcel/Parcelize.wasmJs.kt
@@ -1,3 +1,0 @@
-package org.duchastel.simon.solenne.parcel
-
-actual interface Parcelable
