+++ iosApp/Configuration/Config.xcconfig
@@ -1,3 +1,3 @@
 TEAM_ID=
-BUNDLE_ID=org.duchastel.simon.solenne.SolenneChatApp
+BUNDLE_ID=com.duchastel.simon.solenne.SolenneChatApp
 APP_NAME=SolenneChatApp
